Update project paths and configurations
This commit includes the addition of new vendor paths into the PHP configuration and JavaScript library mappings. Moreover, it extends test configuration in Codeception and Phpspec. The changes also involve the inclusion of new source files in generated files library and removal of certain source folders from the ReturnsDatabase module.