Update store search and data retrieval methods
The methods for search and data retrieval in Store.php have been revised to support additional parameters such as limit, offset, sort column and order. Also, a new count method has been added to return the number of stores. The conversion to array method is removed and replaced with Json serialization. A method to reload data from the database has been added.